+++ src/helm/benchmark/scenarios/vision_language/image2structure/utils_latex.py
